mail/refactor: move IMAP client to mail/ and prefer text/plain over HTML
