+++ dhl_rerouter_poc/mail/email_client_imap.py
@@ -2,7 +2,8 @@
 
 import imaplib
 import email
-from datetime import datetime, timedelta
+from datetime import datetime
+from dhl_rerouter_poc.utils import get_cutoff_since_date
 from .parser import safe_decode, strip_html
 
 import socket
@@ -52,7 +53,7 @@ class ImapEmailClient:
             except Exception as e:
                 logger.error("IMAP login failed for user '%s': %s", self.user, e)
                 return []
-            cutoff = (datetime.now() - timedelta(days=self.lookback)).strftime("%d-%b-%Y")
+            cutoff = get_cutoff_since_date(self.lookback)
             for folder in self.folders:
                 try:
                     status, _ = mail.select(f'"{folder}"', readonly=True)
@@ -71,14 +72,18 @@ class ImapEmailClient:
                         try:
                             _, fetched = mail.fetch(num, "(RFC822)")
                             msg = email.message_from_bytes(fetched[0][1])
-                            body = ""
                             if msg.is_multipart():
+                                text_part = None
+                                html_part = None
                                 for part in msg.walk():
                                     ctype = part.get_content_type()
-                                    if ctype in ("text/plain", "text/html"):
+                                    if ctype == "text/plain" and text_part is None:
                                         ch = part.get_content_charset() or "utf-8"
-                                        txt = safe_decode(part.get_payload(decode=True), ch)
-                                        body += strip_html(txt) if ctype == "text/html" else txt
+                                        text_part = safe_decode(part.get_payload(decode=True), ch)
+                                    elif ctype == "text/html" and html_part is None:
+                                        ch = part.get_content_charset() or "utf-8"
+                                        html_part = strip_html(safe_decode(part.get_payload(decode=True), ch))
+                                body = text_part if text_part is not None else (html_part if html_part is not None else "")
                             else:
                                 ch = msg.get_content_charset() or "utf-8"
                                 body = safe_decode(msg.get_payload(decode=True), ch)
